+++ Example/Example/Data/Model+CohesionKit.swift
@@ -1,14 +1,14 @@
 import CohesionKit
 
 enum Relations {
-    static let matchMarket =
+    static let matchMarkets =
         Relation(
             primaryChildPath: \.match,
-            otherChildren: [.init(\.markets, relation: Relations.marketOutcome)],
+            otherChildren: [.init(\.markets, relation: Relations.marketOutcomes)],
             reduce: { MatchMarkets(match: $0.match, markets: $0.markets) }
         )
     
-    static let marketOutcome =
+    static let marketOutcomes =
         Relation(
             primaryChildPath: \.market,
             otherChildren: [.init(\.outcomes)],