+++ Example/Example/OutcomeButtonStyle.swift
@@ -1,10 +1,3 @@
-//
-//  OutcomeButtonStyle.swift
-//  Example
-//
-//  Created by Jean-Christophe Pastant on 27/05/2021.
-//
-
 import SwiftUI
 
 struct OutcomeButtonStyle: ButtonStyle {